+++ packages/expo-constants/src/Constants.ts
@@ -7,8 +7,6 @@ import type {
   ExpoUpdatesManifest,
   // @ts-ignore -- optional interface, will gracefully degrade to `any` if not installed
 } from 'expo-manifests';
-type DevLauncherManifest = ExpoUpdatesManifest;
-
 import { CodedError, requireOptionalNativeModule } from 'expo-modules-core';
 // @ts-ignore -- optional interface, will gracefully degrade to `any` if not installed
 import type { Manifest as UpdatesManifest, ExpoUpdatesModule } from 'expo-updates';
@@ -26,6 +24,7 @@ import {
   WebManifest,
 } from './Constants.types';
 import ExponentConstants from './ExponentConstants';
+type DevLauncherManifest = ExpoUpdatesManifest;
 
 export {
   AndroidManifest,