+++ packages/expo-image-picker/android/src/main/java/expo/modules/imagepicker/contracts/CropImageContract.kt
@@ -8,7 +8,6 @@ import androidx.core.net.toFile
 import androidx.core.net.toUri
 import androidx.core.os.bundleOf
 import com.canhub.cropper.CropImage
-import com.canhub.cropper.CropImageActivity
 import com.canhub.cropper.CropImageOptions
 import expo.modules.imagepicker.ImagePickerOptions
 import expo.modules.imagepicker.MediaType
@@ -24,7 +23,7 @@ import java.io.Serializable
 internal class CropImageContract(
   private val appContextProvider: AppContextProvider
 ) : AppContextActivityResultContract<CropImageContractOptions, ImagePickerContractResult> {
-  override fun createIntent(context: Context, input: CropImageContractOptions) = Intent(context, CropImageActivity::class.java).apply {
+  override fun createIntent(context: Context, input: CropImageContractOptions) = Intent(context, expo.modules.imagepicker.ExpoCropImageActivity::class.java).apply {
     val mediaType = expo.modules.imagepicker.getType(context.contentResolver, input.sourceUri.toUri())
     val compressFormat = mediaType.toBitmapCompressFormat()
     val cacheDirectory = appContextProvider.appContext.cacheDirectory