+++ packages/expo-router/src/exports.ts
@@ -16,6 +16,7 @@ export {
 export { router, Router } from './imperative-api';
 
 export { Link, Redirect, RedirectProps, LinkProps, WebAnchorProps } from './link/Link';
+export { useIsPreview } from './link/preview/PreviewRouteContext';
 
 export { withLayoutContext } from './layouts/withLayoutContext';
 export { Navigator, Slot };