+++ packages/expo/android/src/main/java/expo/modules/fetch/NativeResponse.kt
@@ -3,6 +3,7 @@
 package expo.modules.fetch
 
 import android.util.Log
+import expo.modules.core.logging.localizedMessageWithCauseLocalizedMessage
 import expo.modules.kotlin.AppContext
 import expo.modules.kotlin.sharedobjects.SharedObject
 import kotlinx.coroutines.CoroutineScope
@@ -68,14 +69,14 @@ internal class NativeResponse(appContext: AppContext, private val coroutineScope
     if (isInvalidState(ResponseState.BODY_STREAMING_STARTED)) {
       return
     }
-    state = ResponseState.BODY_STREAMING_CANCELLED
+    state = ResponseState.BODY_STREAMING_CANCELED
   }
 
-  fun emitRequestCancelled() {
-    val error = FetchRequestCancelledException()
+  fun emitRequestCanceled() {
+    val error = FetchRequestCanceledException()
     this.error = error
     if (state == ResponseState.BODY_STREAMING_STARTED) {
-      emit("didFailWithError", error)
+      emit("didFailWithError", error.localizedMessageWithCauseLocalizedMessage())
     }
     state = ResponseState.ERROR_RECEIVED
   }
@@ -97,7 +98,7 @@ internal class NativeResponse(appContext: AppContext, private val coroutineScope
   //region Callback implementations
 
   override fun onFailure(call: Call, e: IOException) {
-    // Canceled request should be handled by emitRequestCancelled
+    // Canceled request should be handled by emitRequestCanceled
     if (e.message === "Canceled") {
       return
     }
@@ -106,14 +107,14 @@ internal class NativeResponse(appContext: AppContext, private val coroutineScope
         ResponseState.STARTED,
         ResponseState.RESPONSE_RECEIVED,
         ResponseState.BODY_STREAMING_STARTED,
-        ResponseState.BODY_STREAMING_CANCELLED
+        ResponseState.BODY_STREAMING_CANCELED
       )
     ) {
       return
     }
 
     if (state == ResponseState.BODY_STREAMING_STARTED) {
-      emit("didFailWithError", e)
+      emit("didFailWithError", e.localizedMessageWithCauseLocalizedMessage())
     }
     error = e
     state = ResponseState.ERROR_RECEIVED
@@ -174,7 +175,7 @@ internal class NativeResponse(appContext: AppContext, private val coroutineScope
         if (isInvalidState(
             ResponseState.RESPONSE_RECEIVED,
             ResponseState.BODY_STREAMING_STARTED,
-            ResponseState.BODY_STREAMING_CANCELLED
+            ResponseState.BODY_STREAMING_CANCELED
           )
         ) {
           break
@@ -190,7 +191,7 @@ internal class NativeResponse(appContext: AppContext, private val coroutineScope
     } catch (e: IOException) {
       this.error = e
       if (state == ResponseState.BODY_STREAMING_STARTED) {
-        emit("didFailWithError", e)
+        emit("didFailWithError", e.localizedMessageWithCauseLocalizedMessage())
       }
       state = ResponseState.ERROR_RECEIVED
     }